net: macb: fix PHY reset
In newer kernels PHY reset is handled by phylib, but here we don't have
such feature yet.

macb driver tries to handle reset-gpio in PHY DT node, but it fails in
several ways:

- it doesn't correctly handle probe deferring in case GPIO driver needs
  more time to come up

- it doesn't actually reset the PHY, it just make sure it is not in
  reset state

- it doesn't acquire the GPIO, so that it could be exported via sysfs

This patch is a temporary hack to fix those issues; it is expected to
die once we upgrade to newer kernels.