ARM64: zynqmp-uz-eg-som: add NOP usb-2 PHY in order to handle PHY reset GPIO
The USB controller uses two PHYs, one for USB-2 and another for USB-3.

The USB-2 PHY is soldered on the SOM board, so it has to be described in
zynqmp-uz-eg-som.dtsi.

The USB-3 PHY is the internal SoC serdes peripheral, so it's first
described in the SoC dtsi (zynqmp.dtsi); this phy has various possible use
(USB, SATA, ETH) and has various lanes. It's on the carrier board DT (i.e.
zynqmp-uz-eg-iocc.dts) that we declare which serdes lane is routed to the
USB-3 port, that is that we can add the USB-3 PHY to the DWC node.

The USB-2 PHY is the no-xceiv driver (we use it only in order to handle the
PHY reset gpio line), that currently works only with the older USB-PHY
subsystem, so we use usb-phy bindings. (here if we try to use it with
generic PHY bindings we get the DWC driver to finish probe before the PHY
driver has a chance to probe successfully, due to defers waiting for the
GPIO controller driver).

The USB-3 PHY driver works with the newer generic PHY bindings, so we use
it here.

Indeed this casue that PHY-2 and PHY-3 DT stuff are quite splitted up in
two DT file, and they use also different bindings..